# apps/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: PlaidHealth API
  description: |
    Healthcare data aggregation API for connecting to EHR systems and payers.

    ## Authentication

    The API supports two authentication methods:

    1. **API Key** - For B2B integrations, pass your API key in the `X-API-Key` header
    2. **Session** - For user-facing applications using Google OAuth

    ## Rate Limits

    - Widget endpoints: 100 requests/minute
    - FHIR endpoints: 60 requests/minute
    - OAuth endpoints: 10 requests/minute
  version: 1.0.0
  contact:
    email: support@plaidhealth.com

servers:
  - url: https://stripe-healthcare-production.up.railway.app
    description: Production server
  - url: http://localhost:3000
    description: Development server

tags:
  - name: Health
    description: Service health and readiness checks
  - name: Widget
    description: Connect Widget API for patient-initiated connections
  - name: FHIR
    description: FHIR R4 health records API
  - name: Webhooks
    description: Webhook management for real-time notifications
  - name: API Keys
    description: API key management for B2B authentication

paths:
  /health:
    get:
      tags: [Health]
      summary: Basic health check
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time
                  uptime:
                    type: integer
                  version:
                    type: string

  /ready:
    get:
      tags: [Health]
      summary: Readiness check
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [ready]
                  checks:
                    type: object
                    properties:
                      database:
                        type: boolean
                      redis:
                        type: boolean
        '503':
          description: Service not ready

  /api/v1/widget/providers:
    get:
      tags: [Widget]
      summary: List available providers
      responses:
        '200':
          description: List of providers
          content:
            application/json:
              schema:
                type: object
                properties:
                  providers:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        type:
                          type: string
                          enum: [emr, payer, lab]
                        available:
                          type: boolean

  /api/v1/widget/token:
    post:
      tags: [Widget]
      summary: Create widget token
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [client_user_id]
              properties:
                client_user_id:
                  type: string
                redirect_uri:
                  type: string
                  format: uri
                products:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Widget token created
          content:
            application/json:
              schema:
                type: object
                properties:
                  widget_token:
                    type: string
                  expiration:
                    type: string
                    format: date-time
        '400':
          description: Bad request
        '401':
          description: Unauthorized

  /api/v1/widget/initiate/{provider}:
    get:
      tags: [Widget]
      summary: Initiate OAuth flow for a provider
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
          description: Provider ID (e.g., epic, cerner, aetna)
        - name: widget_token
          in: query
          required: true
          schema:
            type: string
          description: Widget token from /widget/token endpoint
      responses:
        '302':
          description: Redirect to provider authorization URL
        '400':
          description: Missing widget_token or invalid provider
        '401':
          description: Invalid or expired widget token

  /api/v1/widget/exchange:
    post:
      tags: [Widget]
      summary: Exchange public token for connection credentials
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [public_token]
              properties:
                public_token:
                  type: string
      responses:
        '200':
          description: Token exchanged successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  connection_id:
                    type: string
                  provider:
                    type: string
                  patient_id:
                    type: string
                  client_user_id:
                    type: string
                  request_id:
                    type: string
        '400':
          description: Missing public_token
        '401':
          description: Invalid or expired public token

  /api/v1/widget/sessions:
    get:
      tags: [Widget]
      summary: Get widget sessions for debugging
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: List of widget sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer

  /api/v1/health-records:
    get:
      tags: [FHIR]
      summary: Get unified health records
      security:
        - ApiKeyAuth: []
      parameters:
        - name: connection_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Health records

  /api/v1/patient:
    get:
      tags: [FHIR]
      summary: Get patient demographics
      security:
        - ApiKeyAuth: []
      parameters:
        - name: connection_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Patient data

  /api/v1/labs:
    get:
      tags: [FHIR]
      summary: Get lab results
      security:
        - ApiKeyAuth: []
      parameters:
        - name: connection_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lab results

  /api/v1/medications:
    get:
      tags: [FHIR]
      summary: Get medications
      security:
        - ApiKeyAuth: []
      parameters:
        - name: connection_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Medications list

  /api/v1/conditions:
    get:
      tags: [FHIR]
      summary: Get conditions/diagnoses
      security:
        - ApiKeyAuth: []
      parameters:
        - name: connection_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Conditions list

  /api/webhooks:
    get:
      tags: [Webhooks]
      summary: List webhooks
      security:
        - SessionAuth: []
      responses:
        '200':
          description: List of webhooks
    post:
      tags: [Webhooks]
      summary: Create webhook
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Webhook created

  /api/keys:
    get:
      tags: [API Keys]
      summary: List API keys
      security:
        - SessionAuth: []
      responses:
        '200':
          description: List of API keys
    post:
      tags: [API Keys]
      summary: Create API key
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                scopes:
                  type: array
                  items:
                    type: string
                    enum: [read, write, admin]
      responses:
        '201':
          description: API key created

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    SessionAuth:
      type: apiKey
      in: cookie
      name: plaidhealth.sid